# Rescan improvements