--- a/src/bls/bls_aggregator.cpp
+++ b/src/bls/bls_aggregator.cpp
@@ -5,6 +5,7 @@
 #include <common/exception.h>
 #include <common/guts.h>
 #include <common/string_util.h>
+#include <common/tracy_shim.h>
 #include <crypto/crypto.h>
 #include <cryptonote_core/cryptonote_core.h>
 #include <l2_tracker/contracts.h>
@@ -13,9 +14,7 @@
 #include <oxenmq/oxenmq.h>
 
 #include <chrono>
-#include <ethyl/utils.hpp>
 #include <memory>
-#include <tracy/Tracy.hpp>
 
 #include "bls_crypto.h"
 #include "crypto/eth.h"