--- a/src/cryptonote_basic/cryptonote_format_utils.h
+++ b/src/cryptonote_basic/cryptonote_format_utils.h
@@ -29,8 +29,8 @@
 // Parts of this file are originally copyright (c) 2012-2013 The Cryptonote developers
 
 #pragma once
-#include <unordered_map>
 #include <tracy/Tracy.hpp>
+#include <unordered_map>
 
 #include "account.h"
 #include "common/guts.h"