TLDR:

Before 

```
[+2h37m31.167s] Loaded subsystems in 9451.04s (snl: 7855.79s; ons: 0.00s; 176.0 blks/s; 603.78 kB/s)
```

After

```
[+53m22.656s] Loaded subsystems in 3201.86s (get blks: 20.01s; snl: 3129.03s; ons: 7.54s; 521.7 blks/s; 38.14 MB/s)
```

Rescan stats over here https://pastebin.com/HUdsgJC5 to avoid the 64k PR description limit.

---

- Put block loading into a thread pool, this only really helps for the first half of the chain in the latter halfs we are stuck doing SQL ops which are slow.
- Removes `data_for_serialization` and instead I've added a specialised function to walk the SNL state and directly serialise it to binary. Before we had `snl -> data for serialization -> binary -> lmdb` now we have `snl -> binary -> lmdb`.
- This includes the changes from https://github.com/oxen-io/oxen-core/pull/1786 which cleans up some tables that are no longer used.
- Propagate `skip_verify` into the SQL DB when rescanning to skip validation of the payments to insert into the SQL DB. With similar rational that if the payments are being reconstructed from rescanning, then they were previously validated when they were added to the DB in the first place.
- Defer updating of the `height` in the SQL DB until you hit a height that is an archive height OR you're in the last window of recent blocks which the table will store for small re-orgs. This code is gated by rescanning, so once you have rescanned to the tip, you will go back to committing data as the blocks are generated.
- Fix off-by-one in the blockchain height check on startup which caused the daemon to pop one block off everytime the daemon started
- Fix `sql_db_has` in SNL that checks if the SQL DB has the required historical state to detach their DBs to from using only the recent table. This fixes the issue where you load up the daemon using the new build and it rescans the _earliest_ archival block it has which for me was block 1'000'000.
- In the SNL, profiling reveals that in multiple functions we were needlessly walking over the entire SNL multiple times for different purposes. As we enter the pulse era and later stages, there were more of these list walks, with 1k+ nodes this starts to get expensive. I've trimmed this down to 2 walks. Initially at the start we walk the list and extract pulse information, expired nodes and the block leader in a single loop. Then after processing the block, walk the list again to extract the active and decomm nodes.

We can take this a step further and collect the delta to SNs in the SNL after processing the block and just apply that to the list. I did try this for a bit but it's quite a risky change to introduce as you need to make sure you cover _all_ the cases where the SNL is modified and make sure they get applied to the list afterwards. So I've backed that out and just walk the SNL before and after the block is processed to reduce the risk of bugs here.

- Similarly, forward the metadata extracted from the SNL into the SQL DB, in this case, the list of payable service nodes.
- For the SNL, another hot spot was the calculating of pulse entropy hashes for seeding the pulse quorum. I've remediated this by having a sliding window of blocks cached in the SNL. As blocks are loaded, we seed this sliding window and incrementally feed blocks in as they are processed. This prevents the expensive look-up into LMDB for 12 historical blocks of data for each block.

The worst of all is still the SQL DB. 2 main culprits, constant writing to the SQL DB and walking 2 very large lists. When you payout SN's, you walk the entire payable list to accumulate rewards into `block_payments`. You then walk the entire `block_payments` list to execute SQL queries.

These are both slow and combine together to make it more than 50% of duration of an update step. I have changes that made the SQL DB cache the changes into runtime structures and commit them periodically. Surprisingly, this only increased the rescanning speed by 2x, not what I was suspecting so there is more things to cut when the release schedule is a bit more stable.

They're actually in this commit tree in the history, kept it there and overwritten it with a commit on top with the current dev changes + re-adding the less risky changes on top. Can revisit if needed in the future.

- This also ships a profiler with the build. Instructions in readme, `-D TRACY_ENABLE=ON` to compile with profiling support and then launch the profiler alongside and `oxend` will stream timings to the app.

At some point, probably before release I will have to look at reducing the memory usage of storing the recent state for reorgs. The current stable build has a long-standing bug where the recent states stored for the SNL only serialise the quorums. Any reorgs that the daemon hit would have made it rescan from the last 10k interval on the stable branch. Fixing it now means that we correctly store the last 360 SNL states which seems to hover around ~500MiB of memory and will get worse as the list grows. 

With stable branch and its serialisation bug, the same 360 SNL states occupied around ~128MiB. This is a fairly big jump in usage that operators may not be happy about. We can look at dropping that down as long as it doesn't affect anything. There's a comment saying that it was used for blink (which I suppose will be deprecated) and it's also related to reorg and checkpointing code that I'll need to verify before dropping it down.

In terms of the 10k history checkpoints, the last changes that were committed capped the nodes to storing the last 2 years worth, so storing those SNL states relatively bounded and won't grow ad-infinitum.

One more thing is that now there's a sizeable time increase in startup times to launch of mainnet. That will also have to be addressed. I believe, the time being spent there is deserializing the large number of SNL states from LMDB on startup.